move maskString in authCommon
