+++ authCommonTs/src/String.ts
@@ -20,25 +20,6 @@ export function makeLower(headers: APIGatewayProxyEventHeaders): APIGatewayProxy
 }
 
 /**
- * Masks a string by showing only the first 2 and last 2 characters.
- * Strings shorter than 6 characters are completely masked.
- *
- * @param stringToMask - The string to mask
- */
-export function maskString(stringToMask: string): string {
-  if (stringToMask.length < 6) {
-    return "".padStart(stringToMask.length, "*");
-  }
-
-  const firstTwoChars = stringToMask.substring(0, 2);
-  const lastTwoChars = stringToMask.substring(stringToMask.length - 2);
-  const hiddenStringLength = stringToMask.length - 4;
-  const hiddenString = "".padStart(hiddenStringLength, "*");
-
-  return firstTwoChars + hiddenString + lastTwoChars;
-}
-
-/**
  * Sanitizes a fiscal number by removing the SPID prefix if present.
  *
  * @param fiscalNumber - The fiscal number to sanitize
