+++ authCommonTs/src/test/String.test.ts
@@ -1,4 +1,4 @@
-import { makeLower, maskString, removeFiscalNumberPrefix, isValidUUID } from "../String";
+import { makeLower, removeFiscalNumberPrefix, isValidUUID } from "../String";
 
 describe("String utils Tests", () => {
   describe("makeLower", () => {
@@ -19,20 +19,6 @@ describe("String utils Tests", () => {
       const result = makeLower({});
 
       expect(result).toEqual({});
-    });
-  });
-
-  describe("maskString", () => {
-    it("should fully mask strings shorter than 6 characters", () => {
-      expect(maskString("abc")).toBe("***");
-      expect(maskString("12345")).toBe("*****");
-      expect(maskString("a")).toBe("*");
-    });
-
-    it("should mask middle characters for strings 6 or more characters long", () => {
-      expect(maskString("123456")).toBe("12**56");
-      expect(maskString("john.doe@example.com")).toBe("jo****************om");
-      expect(maskString("ABCDEFGHIJ")).toBe("AB******IJ");
     });
   });
